Repackage entire project.
This project has not been connected with Orange & Bronze for a while
now, so it makes sense to repackage it to my domain, so as to avoid
any issues later on, particularly when publishing to Maven.
