+++ src/com/jmibanez/tools/jmeter/util/ReflectionUtil.java
@@ -1,0 +1,26 @@
+package com.jmibanez.tools.jmeter.util;
+
+import java.lang.reflect.Field;
+import java.util.Arrays;
+import java.util.ArrayList;
+import java.util.List;
+
+public class ReflectionUtil {
+
+    public static List<Field> getFieldsUpTo(Class<?> startClass,
+                                            Class<?> exclusiveParent) {
+
+        List<Field> currentClassFields =
+            new ArrayList<>(Arrays.asList(startClass.getDeclaredFields()));
+        Class<?> parentClass = startClass.getSuperclass();
+
+        if (parentClass != null &&
+            (exclusiveParent == null || !(parentClass.equals(exclusiveParent)))) {
+            List<Field> parentClassFields =
+                getFieldsUpTo(parentClass, exclusiveParent);
+            currentClassFields.addAll(parentClassFields);
+        }
+
+        return currentClassFields;
+    }
+}
