--- src/com/orangeandbronze/tools/jmeter/MethodRecorder.java
@@ -1,27 +1,0 @@
-/*
- *
- *
- * See README in the source tree for more info
- *
- */
- 
-package com.orangeandbronze.tools.jmeter;
-
-import java.rmi.Remote;
-import java.rmi.RemoteException;
-
-/**
- * Describe interface MethodRecorder here.
- *
- *
- * Created: Fri Nov 14 10:56:53 2008
- *
- * @author <a href="mailto:jm@orangeandbronze.com">JM Ibanez</a>
- * @version 1.0
- */
-public interface MethodRecorder 
-    extends Remote
-{
-    public void recordCall(MethodCallRecord r)
-        throws RemoteException;
-}
